+++ src/__init__.py
@@ -1,0 +1,7 @@
+"""
+CommitGraph - GitHub Contribution Graph Generator
+Modularized version by Shinei Nouzen
+"""
+
+__version__ = "2.0.0"
+__author__ = "Shinei Nouzen"